move security telemetry
