+++ pkg/compliance/agent.go
@@ -18,9 +18,9 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/compliance/k8sconfig"
 	"github.com/DataDog/datadog-agent/pkg/compliance/metrics"
 	"github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/security/common"
 	"github.com/DataDog/datadog-agent/pkg/security/rules"
 	secl "github.com/DataDog/datadog-agent/pkg/security/secl/rules"
+	"github.com/DataDog/datadog-agent/pkg/security/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -63,7 +63,7 @@ type AgentOptions struct {
 type Agent struct {
 	opts AgentOptions
 
-	telemetry  *common.ContainersTelemetry
+	telemetry  *telemetry.ContainersTelemetry
 	statuses   map[string]*CheckStatus
 	statusesMu sync.RWMutex
 
@@ -138,7 +138,7 @@ func NewAgent(opts AgentOptions) *Agent {
 }
 
 func (a *Agent) Start() error {
-	telemetry, err := common.NewContainersTelemetry()
+	telemetry, err := telemetry.NewContainersTelemetry()
 	if err != nil {
 		log.Errorf("could not start containers telemetry: %v", err)
 		return err
