+++ pkg/security/agent/telemetry.go
@@ -11,15 +11,15 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/security/common"
 	"github.com/DataDog/datadog-agent/pkg/security/metrics"
 	"github.com/DataDog/datadog-agent/pkg/security/proto/api"
+	sectelemetry "github.com/DataDog/datadog-agent/pkg/security/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
 // telemetry reports environment information (e.g containers running) when the runtime security component is running
 type telemetry struct {
-	containers            *common.ContainersTelemetry
+	containers            *sectelemetry.ContainersTelemetry
 	runtimeSecurityClient *RuntimeSecurityClient
 	profiledContainers    map[profiledContainer]struct{}
 	logProfiledWorkloads  bool
@@ -31,7 +31,7 @@ func newTelemetry(logProfiledWorkloads, ignoreDDAgentContainers bool) (*telemetr
 		return nil, err
 	}
 
-	containersTelemetry, err := common.NewContainersTelemetry()
+	containersTelemetry, err := sectelemetry.NewContainersTelemetry()
 	if err != nil {
 		return nil, err
 	}
